requirements.txt
# Miniluma 依赖列表

# 基础依赖
aiohttp>=3.8.5
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.4.0
toml>=0.10.2
jsonschema>=4.19.0
typing-extensions>=4.7.1

# LLM 集成
openai>=1.3.0
deepseek>=0.0.2
anthropic>=0.5.0  # Anthropic Claude 支持

# 异步支持
asyncio>=3.4.3

# CLI工具
rich>=13.5.0
colorama>=0.4.6
tqdm>=4.66.1

# 日志和监控
loguru>=0.7.0

# 文件处理
python-magic>=0.4.27; sys_platform != 'win32'
python-magic-bin>=0.4.14; sys_platform == 'win32'

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 网络工具
beautifulsoup4>=4.12.2
lxml>=4.9.3
html2text>=2020.1.16
urllib3>=1.26.18
requests-cache>=1.1.0

# 内容处理
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.2
markdown>=3.4.3
Pillow>=10.0.0

# 代码执行
pyexecjs>=1.5.1

# MCP 集成
playwright>=1.37.0

# 存储
sqlalchemy>=2.0.0
pymongo>=4.5.0
redis>=4.6.0

# 测试工具
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# 代码分析
pylint>=2.17.0
black>=23.7.0
isort>=5.12.0
